⚡️ improve:'BombManager 분리'
